+++ packages/androidx_graphics_shapes/lib/src/material_shapes.dart
@@ -3,6 +3,7 @@ export 'material_shapes/material_shapes.dart'
         RoundedPolygonExtension,
         MorphExtension,
         MaterialShapes,
-        PathBorder,
+        DynamicPathBorder,
+        StaticPathBorder,
         RoundedPolygonBorder,
         MorphBorder;